1.7.x: update to publish to central portal (#1813)
Sonatype is phasing out OSSRH so update the publishing
to go to the new central portal.

Also updates scala, log4j, and spectator to recent versions
and fix various warnings.

+++ atlas-chart/src/main/scala/com/netflix/atlas/chart/graphics/Ticks.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014-2024 Netflix, Inc.
+ * Copyright 2014-2025 Netflix, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -395,7 +395,7 @@ object Ticks {
     if (logDistance <= n) {
       1
     } else {
-      math.ceil(logDistance / n).toInt
+      math.ceil(logDistance.toDouble / n).toInt
     }
   }
